added directory modules into js, moved main to js and the rest js files to modules dir
