+++ challenges.html
@@ -386,9 +386,9 @@
 		  </dialog>
 		<!-- End Modal-dialog -->
 
-		<script type="module" src="main.js"></script>
-		<script type="module"" src="challenges.js"></script>
-		<script src="js/booking.js"></script>
-		<script type="module" src="filter.js"></script>
+		<script type="module" src="js/main.js"></script>
+		<script type="module"" src="js/modules/challenges.js"></script>
+		<script src="js/modules/booking.js"></script>
+		<script type="module" src="js/modules/filter.js"></script>
 	</body>
 </html>
