+++ index.html
@@ -567,7 +567,7 @@
       </div>
     </dialog>
     <!-- End Modal-dialog -->
-    <script type="module" src="main.js"></script>
-    <script src="js/booking.js"></script>
+    <script type="module" src="js/main.js"></script>
+    <script src="js/modules/booking.js"></script>
   </body>
 </html>
