+++ js/modules/challenges.js
